[hotfix][tests] Pull nested test suites out of CheckpointIDCounterTest and follow standard test-base pattern.
That makes the tests consistent with the common practice in Flink and also ensures that the tests are
picked up by the Maven/Surefire build. Previously, these tests were not executed during a Maven build.
